Delete email from users' structure in response for security purposes. Move GetUsersProfile to external file.
