+++ endpoints/users/list.go
@@ -1,0 +1,20 @@
+package users
+
+import (
+	"github.com/bredbrains/tthk-wish-list/database"
+	"github.com/bredbrains/tthk-wish-list/models"
+	"github.com/gin-gonic/gin"
+	"net/http"
+)
+
+func GetUsersProfiles(c *gin.Context) {
+	var users []models.User
+	var err error
+	err, users = database.GetUsers()
+	if err != nil {
+		c.JSON(http.StatusInternalServerError, gin.H{"success": false, "error": err.Error()})
+		return
+	}
+	c.JSON(http.StatusOK, gin.H{"success": true, "users": users})
+	return
+}
